requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-chroma
pypdf
python-dotenv
